requirements.txt
alembic==1.6.2
Brotli==1.0.9
cffi==1.14.5
click==8.0.0
colorama==0.4.4
cryptography==3.4.7
dataclasses==0.8
Flask==1.1.2
Flask-Compress==1.9.0
Flask-Cors==3.0.10
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
greenlet==1.1.0
importlib-metadata==4.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
pycparser==2.20
PyMySQL==1.0.2
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==1.3.18
SQLAlchemy-Utils==0.36.8
typing-extensions==3.10.0.0
Werkzeug==2.0.1
WTForms==2.2.1
zipp==3.4.1

pandas~=1.1.5